Allow env var config, rename to defaults.js
