+++ src/commandManager.js
@@ -1,4 +1,4 @@
-import config from '../config.js';
+import config from './config.js';
 
 import fs from 'fs';
 import url from 'url';
